Chapter06 사용자 인증
- 사용자 인증 (로그인, 로그아웃) 구현

+++ Chapter06/src/main/java/edu/pnu/controller/BoardController.java
@@ -6,12 +6,16 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.SessionAttributes;
 
 import edu.pnu.domain.Board;
+import edu.pnu.domain.Member;
 import edu.pnu.service.BoardService;
 
+@SessionAttributes("member")
 @Controller
 public class BoardController {
 	@GetMapping("/hello")
@@ -22,10 +26,16 @@ public class BoardController {
 	@Autowired
 	private BoardService boardService;
 	
-	
+	@ModelAttribute("member") // 구동하자마자 세션에 빈 객체를 넣기 위해
+	public Member setMember() {
+		return new Member();
+	}
 	
 	@RequestMapping("/getBoardList")
-	public String getBoardList(Model model) {
+	public String getBoardList(@ModelAttribute("member") Member member, Model model) {
+		if(member.getId()==null) {
+			return "redirect:login";
+		}
 		List<Board> boardList = boardService.getBoardList();
 		
 		model.addAttribute("boardList",boardList);
@@ -33,30 +43,46 @@ public class BoardController {
 	}
 	
 	@GetMapping("/insertBoard") //insert 화면을 띄워주기 위해 필요
-	public String insertBoardView() {
+	public String insertBoardView(@ModelAttribute("member") Member member) {
+		if(member.getId()==null) {
+			return "redirect:login";
+		}
+		
 		return "insertBoard";
 	}
 
 	@PostMapping("/insertBoard") //내용을 저장하고 리다이렉트를 위해 필요
-	public String insertBoard(Board board) {
+	public String insertBoard(@ModelAttribute("member") Member member,Board board) {
+		if(member.getId()==null) {
+			return "redirect:login";
+		}
 		boardService.insertBoard(board);
 		return "redirect:getBoardList";
 	}
 	
 	@GetMapping("/getBoard")
-	public String getBoard(Board board, Model model) {
+	public String getBoard(@ModelAttribute("member") Member member,Board board, Model model) {
+		if(member.getId()==null) {
+			return "redirect:login";
+		}
 		model.addAttribute("board", boardService.getBoard(board));
 		return "getBoard";
 	}
 	
 	@PostMapping("/updateBoard")
-	public String updateBoard(Board board) {
+	public String updateBoard(@ModelAttribute("member") Member member,Board board) {
+		if(member.getId()==null) {
+			return "redirect:login";
+		}
 		boardService.updateBoard(board);
 		return "forward:getBoardList";
 	}
 	
 	@GetMapping("/deleteBoard")
-	public String deleteBoard(Board board) {
+	public String deleteBoard(@ModelAttribute("member") Member member,Board board) {
+		if(member.getId()==null) {
+			return "redirect:login";
+		}
 		boardService.deleteBoard(board);
 		return "forward:getBoardList";
 	}
